# openapi/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: User Management API (Spring Boot)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  description: |
    Backed SprinBoot Application API supporting Angular application for User Authentication and CRUD operation.
servers:
  - url: https://{tenant}/api/v1
    description: Tenant server
    variables:
      tenant:
        default: www
        description: Your tenant id
  - url: https://example.com/api/v1
    description: Example server
tags:
  - name: Authentication
    description: Authentication operations.
  - name: Admin
    description: Administrative user management operations.
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User authentication
      description: This operation allows users to authenticate with the system. The user credentials are passed in the request body.
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest_400'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '403':
          $ref: '#/components/responses/Forbidden_403'
        '500':
          $ref: '#/components/responses/InternalServerError_500'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: This operation allows users to logout of the system.
      operationId: logoutUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CookieParam'
      responses:
        '204':
          description: Logout successful
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetCookie'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '500':
          $ref: '#/components/responses/InternalServerError_500'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh user token
      description: Refreshes an expired or expiring authentication token using the refresh token from the Set-Cookie header.
      operationId: refreshToken
      security: []
      parameters:
        - $ref: '#/components/parameters/CookieParam'
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest_400'
        '401':
          $ref: '#/components/responses/Unauthorized_401'
        '500':
          $ref: '#/components/responses/InternalServerError_500'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: This operation allows the creation of a new user in the system.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        description: Data required to register a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: Successfully registered a new user
        '405':
          description: Invalid input
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get list of users
      description: This operation retrieves a list of all users in the system.
      operationId: getUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
  /admin/users/{id}:
    get:
      tags:
        - Admin
      summary: Get user by id
      description: This operation retrieves the user profile for the specified user.
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: The id of the user profile to be retrieved.
          example: 101
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid username provided
          $ref: '#/components/responses/BadRequest_400'
        '403':
          description: Forbidden
          $ref: '#/components/responses/Problem'
        '404':
          description: User not found
          $ref: '#/components/responses/NotFound_404'
    put:
      tags:
        - Admin
      summary: Update user data
      description: This operation allows users to update the information in their own profile. The update can only be performed by the logged in user.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID of the user for whom to update the profile.
          example: 101
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        description: Updated user object
        required: true
      responses:
        '200':
          description: Successfully updated the user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid username provided
          $ref: '#/components/responses/BadRequest_400'
        '404':
          description: User not found
          $ref: '#/components/responses/NotFound_404'
        '405':
          description: Method not allowed
          content:
            application/json:
              example:
                status: 405
                message: Not allowed
    delete:
      tags:
        - Admin
      summary: Delete a user
      description: This operation deletes all data associated with the requested `username`. Only administrators can perform this operation.
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: The name of the user profile to be **deleted**.
          example: 101
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successfully deleted a user profile
        '400':
          description: Invalid username provided
          $ref: '#/components/responses/BadRequest_400'
        '404':
          description: User not found
          $ref: '#/components/responses/NotFound_404'
  /admin/users/{id}/roles:
    put:
      tags:
        - Admin
      summary: Assign roles to a user
      description: This operation allows users to assign roles to themselves. The roles can be assigned to the user by specifying the role name in the request body.
      operationId: assignRolesToUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID of the user for whom to assign roles.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoles'
        description: Updated user object
        required: true
      responses:
        '200':
          description: Successfully updated the user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid username provided
          $ref: '#/components/responses/BadRequest_400'
        '404':
          description: User not found
          $ref: '#/components/responses/NotFound_404'
        '405':
          description: Method not allowed
          content:
            application/json:
              example:
                status: 405
                message: Not allowed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[a-zA-Z0-9._@-]+$
          description: User's email or username
          example: johndoe
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: User's password (minimum 8 characters)
          example: strongpassword123
    AuthTokenResponse:
      type: object
      required:
        - token
        - type
        - expiresIn
        - expiresAt
      properties:
        token:
          type: string
          description: The authentication token (JWT).
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Auth token lifetime in seconds.
          example: 3600
        expiresAt:
          type: string
          format: date-time
          description: Auth token expiration timestamp.
          example: '2026-01-29T12:30:00Z'
    UserRole:
      type: string
      description: The role of the user.
      example: USER_ROLE
      enum:
        - USER_ROLE
        - ACCOUNT_ADMIN
        - DOCUMENT_ADMIN
        - ITEM_ADMIN
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the user.
        username:
          type: string
          description: The username of the user.
        firstName:
          type: string
          description: The first name of the user.
        lastName:
          type: string
          description: The last name of the user.
        email:
          type: string
          description: The email address of the user.
        enabled:
          type: boolean
          description: A flag indicating whether the user is enabled or not.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/AuthTokenResponse'
        - type: object
          required:
            - user
      properties:
        refreshTokenExpiresAt:
          type: string
          format: date-time
          description: The expiration date of the refresh token (sent in Set-Cookie header).
          example: '2026-02-05T10:30:00Z'
        user:
          $ref: '#/components/schemas/UserResponse'
          description: The user object.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
    Problem:
      type: object
      additionalProperties: true
      minProperties: 1
      description: The Problem Details JSON Object [[RFC7807](https://tools.ietf.org/html/rfc7807)].
      properties:
        type:
          type: string
          description: A URI reference [[RFC3986](https://tools.ietf.org/html/rfc3986)] that identifies the problem type. It should provide human-readable documentation for the problem type. When this member is not present, its value is assumed to be "about:blank".
          format: uri
        title:
          type: string
          description: A short, human-readable summary of the problem type. It SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization.
        status:
          type: integer
          description: The HTTP status code.
          minimum: 400
          maximum: 599
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem.  It may or may not yield further information if dereferenced.
        timestamp:
          type: string
          format: date-time
          description: The date and time when the problem occurred. It MUST be in UTC time.
    UserRegistration:
      type: object
      required:
        - firstName
        - lastName
        - username
        - password
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
          format: password
    UserUpdate:
      type: object
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        enabled:
          type: boolean
          default: true
    UserRoles:
      type: object
      properties:
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/UserRole'
          description: An array of roles for the user.
  headers:
    SetCookie:
      description: Sets or clears the HTTP refresh token cookie
      schema:
        type: string
      examples:
        login:
          summary: Login response - sets refresh token cookie (7 days)
          value: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; Path=/auth/refresh; HttpOnly; Secure; SameSite=Strict; Max-Age=604800
        logout:
          summary: Logout response - clears refresh token cookie
          value: refreshToken=; Path=/auth/refresh; HttpOnly; Secure; Max-Age=0
  responses:
    BadRequest_400:
      description: Bad request due to invalid input
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized_401:
      description: Unauthorized - Invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden_403:
      description: Forbidden - User is not authorized to use this resource
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError_500:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Problem:
      description: Problem
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound_404:
      description: The requested resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  parameters:
    CookieParam:
      name: Cookie
      in: header
      description: Refresh token cookie (declared for generator/backend). Clients normally send Cookie header; backend expects this header declared.
      required: true
      schema:
        type: string
        example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; Path=/auth/refresh; HttpOnly; Secure; SameSite=Strict; Max-Age=604800
x-tagGroups:
  - name: General
    tags:
      - Authentication
      - Admin